+++ test/unit_tests/datautil/test_serialization.py
@@ -11,7 +11,7 @@ import pandas as pd
 
 from braindecode.datasets import BaseConcatDataset, MOABBDataset
 from braindecode.preprocessing import (
-    create_windows_from_events, Preprocessor, preprocess)
+    create_windows_from_events, Preprocessor, preprocess,)
 from braindecode.datautil.serialization import (
     load_concat_dataset, _check_save_dir_empty)
 
@@ -30,6 +30,16 @@ def setup_concat_windows_dataset(setup_concat_raw_dataset):
         trial_stop_offset_samples=0)
 
 
+@pytest.fixture()
+def setup_concat_windows_mne_dataset(setup_concat_raw_dataset):
+    moabb_dataset = setup_concat_raw_dataset
+    return create_windows_from_events(
+        concat_ds=moabb_dataset,
+        trial_start_offset_samples=0,
+        trial_stop_offset_samples=0,
+        picks=moabb_dataset.datasets[0].raw.ch_names)
+
+
 def test_outdated_save_concat_raw_dataset(setup_concat_raw_dataset, tmpdir):
     concat_raw_dataset = setup_concat_raw_dataset
     n_raw_datasets = len(concat_raw_dataset.datasets)
@@ -44,8 +54,8 @@ def test_outdated_save_concat_raw_dataset(setup_concat_raw_dataset, tmpdir):
 
 
 def test_outdated_save_concat_windows_dataset(
-        setup_concat_windows_dataset, tmpdir):
-    concat_windows_dataset = setup_concat_windows_dataset
+        setup_concat_windows_mne_dataset, tmpdir):
+    concat_windows_dataset = setup_concat_windows_mne_dataset
     n_windows_datasets = len(concat_windows_dataset.datasets)
     with pytest.warns(
             UserWarning, match='This function only exists for '
@@ -60,16 +70,9 @@ def test_outdated_save_concat_windows_dataset(
 def test_load_concat_raw_dataset(setup_concat_raw_dataset, tmpdir):
     concat_raw_dataset = setup_concat_raw_dataset
     n_raw_datasets = len(concat_raw_dataset.datasets)
-    with pytest.warns(
-            UserWarning, match='This function only exists for '
-                               'backwards compatibility purposes. DO NOT USE!'):
-        concat_raw_dataset._outdated_save(path=tmpdir, overwrite=False)
-    with pytest.warns(
-            UserWarning, match="The way your dataset was saved is deprecated by"
-                               " now. Please save it again using dataset.save()"
-                               "."):
-        loaded_concat_raw_dataset = load_concat_dataset(
-            path=tmpdir, preload=False)
+    concat_raw_dataset.save(path=tmpdir, overwrite=False)
+    loaded_concat_raw_dataset = load_concat_dataset(
+        path=tmpdir, preload=False)
     assert len(concat_raw_dataset) == len(loaded_concat_raw_dataset)
     assert (len(concat_raw_dataset.datasets) ==
             len(loaded_concat_raw_dataset.datasets))
@@ -83,18 +86,18 @@ def test_load_concat_raw_dataset(setup_concat_raw_dataset, tmpdir):
         concat_raw_dataset.description, loaded_concat_raw_dataset.description)
 
 
+def test_fail_outdated_save_on_new_window_dataset(setup_concat_windows_dataset, tmpdir):
+    concat_windows_dataset = setup_concat_windows_dataset
+    with pytest.raises(
+            NotImplementedError, match='Outdated save not implemented for new window datasets.'):
+        concat_windows_dataset._outdated_save(path=tmpdir, overwrite=False)
+
+
 def test_load_concat_windows_dataset(setup_concat_windows_dataset, tmpdir):
     concat_windows_dataset = setup_concat_windows_dataset
     n_windows_datasets = len(concat_windows_dataset.datasets)
-    with pytest.warns(
-            UserWarning, match='This function only exists for '
-                               'backwards compatibility purposes. DO NOT USE!'):
-        concat_windows_dataset._outdated_save(path=tmpdir, overwrite=False)
-    with pytest.warns(
-            UserWarning, match="The way your dataset was saved is deprecated by"
-                               " now. Please save it again using dataset.save()"
-                               "."):
-        loaded_concat_windows_dataset = load_concat_dataset(
+    concat_windows_dataset.save(path=tmpdir, overwrite=False)
+    loaded_concat_windows_dataset = load_concat_dataset(
             path=tmpdir, preload=False)
     assert len(concat_windows_dataset) == len(loaded_concat_windows_dataset)
     assert (len(concat_windows_dataset.datasets) ==
@@ -114,19 +117,10 @@ def test_load_concat_windows_dataset(setup_concat_windows_dataset, tmpdir):
 def test_load_multiple_concat_raw_dataset(setup_concat_raw_dataset, tmpdir):
     concat_raw_dataset = setup_concat_raw_dataset
     for i in range(2):
-        path = os.path.join(tmpdir, str(i))
-        os.makedirs(path)
-        with pytest.warns(
-                UserWarning, match='This function only exists for '
-                                   'backwards compatibility purposes. DO NOT '
-                                   'USE!'):
-            concat_raw_dataset._outdated_save(path=path, overwrite=False)
-        with pytest.warns(
-                UserWarning, match="The way your dataset was saved is "
-                                   "deprecated by now. Please save it again "
-                                   "using dataset.save()."):
-            loaded_concat_raw_datasets = load_concat_dataset(
-                path=tmpdir, preload=False)
+        i_offset = i * len(concat_raw_dataset.datasets)
+        concat_raw_dataset.save(path=tmpdir, overwrite=False, offset=i_offset)
+    loaded_concat_raw_datasets = load_concat_dataset(
+        path=tmpdir, preload=False)
     assert 2 * len(concat_raw_dataset) == len(loaded_concat_raw_datasets)
     assert (2 * len(concat_raw_dataset.datasets) ==
             len(loaded_concat_raw_datasets.datasets))
@@ -134,23 +128,13 @@ def test_load_multiple_concat_raw_dataset(setup_concat_raw_dataset, tmpdir):
             len(loaded_concat_raw_datasets.description))
 
 
-def test_load_multiple_concat_windows_dataset(setup_concat_windows_dataset,
-                                              tmpdir):
+def test_load_multiple_concat_windows_dataset(setup_concat_windows_dataset, tmpdir):
     concat_windows_dataset = setup_concat_windows_dataset
     for i in range(2):
-        path = os.path.join(tmpdir, str(i))
-        os.makedirs(path)
-        with pytest.warns(
-                UserWarning, match='This function only exists for '
-                                   'backwards compatibility purposes. DO NOT '
-                                   'USE!'):
-            concat_windows_dataset._outdated_save(path=path, overwrite=False)
-        with pytest.warns(
-                UserWarning, match="The way your dataset was saved is "
-                                   "deprecated by now. Please save it again "
-                                   "using dataset.save()."):
-            loaded_concat_windows_datasets = load_concat_dataset(
-                path=tmpdir, preload=False)
+        i_offset = i * len(concat_windows_dataset.datasets)
+        concat_windows_dataset.save(path=tmpdir, overwrite=False, offset=i_offset)
+    loaded_concat_windows_datasets = load_concat_dataset(
+        path=tmpdir, preload=False)
     assert 2 * len(concat_windows_dataset) == len(loaded_concat_windows_datasets)
     assert (2 * len(concat_windows_dataset.datasets) ==
             len(loaded_concat_windows_datasets.datasets))
@@ -187,7 +171,7 @@ def test_load_save_window_preproc_kwargs(setup_concat_windows_dataset, tmpdir):
     for i in range(len(concat_windows_dataset.datasets)):
         subdir = os.path.join(tmpdir, str(i))
         assert os.path.exists(os.path.join(subdir, 'window_kwargs.json'))
-        assert os.path.exists(os.path.join(subdir, 'window_preproc_kwargs.json'))
+        assert os.path.exists(os.path.join(subdir, 'raw_preproc_kwargs.json'))
     loaded_concat_windows_dataset = load_concat_dataset(tmpdir, preload=False)
 
     for ds in loaded_concat_windows_dataset.datasets:
@@ -198,11 +182,11 @@ def test_load_save_window_preproc_kwargs(setup_concat_windows_dataset, tmpdir):
                 'window_size_samples': None, 'window_stride_samples': None,
                 'drop_last_window': False, 'mapping':  {
                     'feet': 0, 'left_hand': 1, 'right_hand': 2, 'tongue': 3},
-                'preload': False, 'drop_bad_windows': True, 'picks': None,
+                'preload': False, 'drop_bad_windows': None, 'picks': None,
                 'reject': None, 'flat': None, 'on_missing': 'error',
-                'accepted_bads_ratio': 0.0, 'verbose': 'error'})
+                'accepted_bads_ratio': 0.0, 'verbose': 'error', 'use_mne_epochs': False})
         ]
-        assert ds.window_preproc_kwargs == [
+        assert ds.raw_preproc_kwargs == [
             ('pick_channels', {'ch_names': ['Cz']}),
         ]
 
@@ -231,7 +215,7 @@ def test_save_concat_windows_dataset(setup_concat_windows_dataset, tmpdir):
     for windows_i in range(n_windows_datasets):
         subdir = os.path.join(tmpdir, str(windows_i))
         assert os.path.exists(os.path.join(subdir, "description.json"))
-        assert os.path.exists(os.path.join(subdir, f"{windows_i}-epo.fif"))
+        assert os.path.exists(os.path.join(subdir, f"{windows_i}-raw.fif"))
     assert not os.path.exists(os.path.join(tmpdir, f"{n_windows_datasets}"))
 
 
@@ -270,10 +254,8 @@ def test_load_concat_windows_dataset_parallel(setup_concat_windows_dataset, tmpd
     with pytest.warns(None) as raised_warnings:
         concat_windows_dataset.save(path=tmpdir, overwrite=False)
         assert len(raised_warnings) == 0
-    # assert warning raised because of n_jobs not supported with mne.Epochs
-    with pytest.warns(UserWarning, match='Parallelized reading with '
-                                         '`preload=False` is not supported for '
-                                         'windowed data. Will use `n_jobs=1`.'):
+    # assert no warning raised anymore as underlying data is raw now
+    with pytest.warns(None) as raised_warnings:
         loaded_concat_windows_dataset = load_concat_dataset(
             path=tmpdir, preload=False, n_jobs=2)
         assert len(raised_warnings) == 0